refactor: split actress application into query and maintenance seams
Replace the wide actressApplicationService façade with policy-bearing query/maintenance modules and wire conflict IPC directly to the identity workflow.
