+++ scripts/check-actress-boundaries.mjs
@@ -121,13 +121,19 @@ for (const file of sourceFiles('src/renderer/src')) {
 }
 
 const actressHandler = 'src/main/ipc/actressHandlers.ts'
+const actressApplicationSeams = new Set([
+  '../services/actressQueryService',
+  '../services/actressMaintenanceService',
+  '../services/actressIdentityConflictWorkflow'
+])
 for (const specifier of importsOf(actressHandler)) {
   if (
     /\.\.\/(db|scrapers)\//.test(specifier) ||
-    (/\.\.\/services\//.test(specifier) &&
-      specifier !== '../services/actressApplicationService')
+    (/\.\.\/services\//.test(specifier) && !actressApplicationSeams.has(specifier))
   ) {
-    violations.push(`${actressHandler}: actress IPC must delegate through its application service`)
+    violations.push(
+      `${actressHandler}: actress IPC must delegate through query / maintenance / identity seams`
+    )
   }
 }
 
@@ -172,7 +178,8 @@ for (const file of sourceFiles('src/main/db')) {
 
 for (const file of [
   'src/main/services/videoService.ts',
-  'src/main/services/actressApplicationService.ts',
+  'src/main/services/actressQueryService.ts',
+  'src/main/services/actressMaintenanceService.ts',
   'src/main/services/actressGalleryService.ts',
   'src/main/services/actressIdentityConflictWorkflow.ts',
   'src/main/scrapers/actressScraperManager.ts',
